Implement custom deserializer for Store structs

+++ src/store.rs
@@ -5,11 +5,11 @@ use super::redis::{self, Commands, RedisResult};
 #[derive(Clone)]
 pub struct Store {
     pub client: redis::Client,
-    pub expire_keys: usize, // Redis key TTL, in seconds
+    pub expire_keys: u64, // Redis key TTL, in seconds
 }
 
 impl Store {
-    pub fn new(url: &str, expire_keys: usize) -> Result<Store, &'static str> {
+    pub fn new(url: &str, expire_keys: u64) -> Result<Store, &'static str> {
         let res = redis::Client::open(url);
         if res.is_err() {
             return Err("error opening store connection");
@@ -23,7 +23,8 @@ impl Store {
         if res.is_err() {
             return Err(res.unwrap_err().to_string());
         }
-        let res: RedisResult<bool> = self.client.expire(key, self.expire_keys);
+        let res: RedisResult<bool> =
+            self.client.expire(key, self.expire_keys as usize);
         if res.is_err() {
             return Err(res.unwrap_err().to_string());
         }
